re-organize tests into standard format
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/GSEABase@104608 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/methods-OBOCollection.R
@@ -189,7 +189,7 @@ setAs("OBOCollection", "graphNEL",
           s <- .stanza(from)
           k <- .kv(from)
           nodes <- .OBOids(s)
-          edgeL <- rep(list(list(edges=numeric())), length(nodes))
+          edgeL <- rep(list(list(edges=integer())), length(nodes))
           names(edgeL) <- nodes
           df <- k[k$key=="is_a" & k$value %in% nodes,
                   !names(k)=="key",drop=FALSE]
